+++ src/compiler/compile/nodes/Window.ts
@@ -37,7 +37,7 @@ export default class Window extends Node {
 					const { parts } = flatten_reference(node.expression);
 
 					// TODO is this constraint necessary?
-					component.error(node.expression, compiler_errors.invalid_binding_window(parts));
+					return component.error(node.expression, compiler_errors.invalid_binding_window(parts));
 				}
 
 				if (!~valid_bindings.indexOf(node.name)) {
@@ -48,9 +48,9 @@ export default class Window extends Node {
 					);
 
 					if (match) {
-						component.error(node, compiler_errors.invalid_binding_on(node.name, '<svelte:window>', ` (did you mean '${match}'?)`));
+						return component.error(node, compiler_errors.invalid_binding_on(node.name, '<svelte:window>', ` (did you mean '${match}'?)`));
 					} else {
-						component.error(node, compiler_errors.invalid_binding_on(node.name, '<svelte:window>', ` — valid bindings are ${list(valid_bindings)}`));
+						return component.error(node, compiler_errors.invalid_binding_on(node.name, '<svelte:window>', ` — valid bindings are ${list(valid_bindings)}`));
 					}
 				}