+++ src/compiler/interfaces.ts
@@ -24,6 +24,12 @@ export interface MustacheTag extends BaseNode {
 	expression: Node;
 }
 
+export interface Comment extends BaseNode {
+	type: 'Comment';
+	data: string;
+	ignores: string[];
+}
+
 export type DirectiveType = 'Action'
 | 'Animation'
 | 'Binding'
@@ -53,7 +59,8 @@ export type TemplateNode = Text
 | MustacheTag
 | BaseNode
 | Directive
-| Transition;
+| Transition
+| Comment;
 
 export interface Parser {
 	readonly template: string;
@@ -117,6 +124,7 @@ export interface CompileOptions {
 	name?: string;
 	filename?: string;
 	generate?: 'dom' | 'ssr' | false;
+	errorMode?: 'throw' | 'warn';
 	varsReport?: 'full' | 'strict' | false;
 
 	sourcemap?: object | string;