+++ src/runtime/index.ts
@@ -7,6 +7,7 @@ export {
 	afterUpdate,
 	setContext,
 	getContext,
+	getAllContexts,
 	hasContext,
 	tick,
 	createEventDispatcher,