+++ src/runtime/internal/dom.ts
@@ -14,7 +14,7 @@ export function end_hydrating() {
 type NodeEx = Node & {
 	claim_order?: number,
 	hydrate_init? : true,
-	actual_end_child?: Node,
+	actual_end_child?: NodeEx,
 	childNodes: NodeListOf<NodeEx>,
 };
 
@@ -37,8 +37,20 @@ function init_hydrate(target: NodeEx) {
 
 	type NodeEx2 = NodeEx & {claim_order: number};
 
-	// We know that all children have claim_order values since the unclaimed have been detached
-	const children = target.childNodes as NodeListOf<NodeEx2>;
+	// We know that all children have claim_order values since the unclaimed have been detached if target is not <head>
+	let children: ArrayLike<NodeEx2> = target.childNodes as NodeListOf<NodeEx2>;
+
+	// If target is <head>, there may be children without claim_order
+	if (target.nodeName === 'HEAD') {
+		const myChildren = [];
+		for (let i = 0; i < children.length; i++) {
+			const node = children[i];
+			if (node.claim_order !== undefined) {
+				myChildren.push(node);
+			}
+		}
+		children = myChildren;
+	}
 
 	/*
 	* Reorder claimed children optimally.
@@ -70,7 +82,8 @@ function init_hydrate(target: NodeEx) {
 		// Find the largest subsequence length such that it ends in a value less than our current value
 
 		// upper_bound returns first greater value, so we subtract one
-		const seqLen = upper_bound(1, longest + 1, idx => children[m[idx]].claim_order, current) - 1;
+		// with fast path for when we are on the current longest subsequence
+		const seqLen = ((longest > 0 && children[m[longest]].claim_order <= current) ? longest + 1 : upper_bound(1, longest, idx => children[m[idx]].claim_order, current)) - 1;
 
 		p[i] = m[seqLen] + 1;
 
@@ -112,15 +125,63 @@ function init_hydrate(target: NodeEx) {
 	}
 }
 
-export function append(target: NodeEx, node: NodeEx) {
+export function append(target: Node, node: Node) {
+	target.appendChild(node);
+}
+
+export function append_styles(
+	target: Node,
+	style_sheet_id: string,
+	styles: string
+) {
+	const append_styles_to = get_root_for_style(target);
+
+	if (!append_styles_to.getElementById(style_sheet_id)) {
+		const style = element('style');
+		style.id = style_sheet_id;
+		style.textContent = styles;
+		append_stylesheet(append_styles_to, style);
+	}
+}
+
+export function get_root_for_style(node: Node): ShadowRoot | Document {
+	if (!node) return document;
+
+	const root = node.getRootNode ? node.getRootNode() : node.ownerDocument;
+	if ((root as ShadowRoot).host) {
+		return root as ShadowRoot;
+	}
+	return document;
+}
+
+export function append_empty_stylesheet(node: Node) {
+	const style_element = element('style') as HTMLStyleElement;
+	append_stylesheet(get_root_for_style(node), style_element);
+	return style_element;
+}
+
+function append_stylesheet(node: ShadowRoot | Document, style: HTMLStyleElement) {
+	append((node as Document).head || node, style);
+}
+
+export function append_hydration(target: NodeEx, node: NodeEx) {
 	if (is_hydrating) {
 		init_hydrate(target);
 
 		if ((target.actual_end_child === undefined) || ((target.actual_end_child !== null) && (target.actual_end_child.parentElement !== target))) {
 			target.actual_end_child = target.firstChild;
 		}
+
+		// Skip nodes of undefined ordering
+		while ((target.actual_end_child !== null) && (target.actual_end_child.claim_order === undefined)) {
+			target.actual_end_child = target.actual_end_child.nextSibling;
+		}
+
 		if (node !== target.actual_end_child) {
-			target.insertBefore(node, target.actual_end_child);
+			// We only insert if the ordering of this node should be modified or the parent node is not target
+			if (node.claim_order !== undefined || node.parentNode !== target) {
+				target.insertBefore(node, target.actual_end_child);
+			}
 		} else {
 			target.actual_end_child = node.nextSibling;
 		}
@@ -129,9 +190,13 @@ export function append(target: NodeEx, node: NodeEx) {
 	}
 }
 
-export function insert(target: NodeEx, node: NodeEx, anchor?: NodeEx) {
+export function insert(target: Node, node: Node, anchor?: Node) {
+	target.insertBefore(node, anchor || null);
+}
+
+export function insert_hydration(target: NodeEx, node: NodeEx, anchor?: NodeEx) {
 	if (is_hydrating && !anchor) {
-		append(target, node);
+		append_hydration(target, node);
 	} else if (node.parentNode !== target || node.nextSibling != anchor) {
 		target.insertBefore(node, anchor || null);
 	}
@@ -304,11 +369,15 @@ export function children(element: Element) {
 	return Array.from(element.childNodes);
 }
 
-function claim_node<R extends ChildNodeEx>(nodes: ChildNodeArray, predicate: (node: ChildNodeEx) => node is R, processNode: (node: ChildNodeEx) => void, createNode: () => R, dontUpdateLastIndex: boolean = false) {
-	// Try to find nodes in an order such that we lengthen the longest increasing subsequence
+function init_claim_info(nodes: ChildNodeArray) {
 	if (nodes.claim_info === undefined) {
 		nodes.claim_info = {last_index: 0, total_claimed: 0};
 	}
+}
+
+function claim_node<R extends ChildNodeEx>(nodes: ChildNodeArray, predicate: (node: ChildNodeEx) => node is R, processNode: (node: ChildNodeEx) => ChildNodeEx | undefined, createNode: () => R, dontUpdateLastIndex: boolean = false) {
+	// Try to find nodes in an order such that we lengthen the longest increasing subsequence
+	init_claim_info(nodes);
 
 	const resultNode = (() => {
 		// We first try to find an element after the previous one
@@ -316,9 +385,13 @@ function claim_node<R extends ChildNodeEx>(nodes: ChildNodeArray, predicate: (no
 			const node = nodes[i];
 
 			if (predicate(node)) {
-				processNode(node);
+				const replacement = processNode(node);
 
-				nodes.splice(i, 1);
+				if (replacement === undefined) {
+					nodes.splice(i, 1);
+				} else {
+					nodes[i] = replacement;
+				}
 				if (!dontUpdateLastIndex) {
 					nodes.claim_info.last_index = i;
 				}
@@ -333,12 +406,16 @@ function claim_node<R extends ChildNodeEx>(nodes: ChildNodeArray, predicate: (no
 			const node = nodes[i];
 
 			if (predicate(node)) {
-				processNode(node);
+				const replacement = processNode(node);
 
-				nodes.splice(i, 1);
+				if (replacement === undefined) {
+					nodes.splice(i, 1);
+				} else {
+					nodes[i] = replacement;
+				}
 				if (!dontUpdateLastIndex) {
 					nodes.claim_info.last_index = i;
-				} else {
+				} else if (replacement === undefined) {
 					// Since we spliced before the last_index, we decrease it
 					nodes.claim_info.last_index--;
 				}
@@ -368,6 +445,7 @@ export function claim_element(nodes: ChildNodeArray, name: string, attributes: {
 				}
 			}
 			remove.forEach(v => node.removeAttribute(v));
+			return undefined;
 		},
 		() => svg ? svg_element(name as keyof SVGElementTagNameMap) : element(name as keyof HTMLElementTagNameMap)
 	);
@@ -378,7 +456,14 @@ export function claim_text(nodes: ChildNodeArray, data) {
 		nodes,
 		(node: ChildNode): node is Text => node.nodeType === 3,
 		(node: Text) => {
-			node.data = '' + data;
+			const dataStr = '' + data;
+			if (node.data.startsWith(dataStr)) {
+				if (node.data.length !== dataStr.length) {
+					return node.splitText(dataStr.length);
+				}
+			} else {
+				node.data = dataStr;
+			}
 		},
 		() => text(data),
 		true	// Text nodes should not update last index since it is likely not worth it to eliminate an increasing subsequence of actual elements
@@ -404,12 +489,19 @@ export function claim_html_tag(nodes) {
 	const start_index = find_comment(nodes, 'HTML_TAG_START', 0);
 	const end_index = find_comment(nodes, 'HTML_TAG_END', start_index);
 	if (start_index === end_index) {
-		return new HtmlTag();
+		return new HtmlTagHydration();
 	}
+
+	init_claim_info(nodes);
 	const html_tag_nodes = nodes.splice(start_index, end_index + 1);
 	detach(html_tag_nodes[0]);
 	detach(html_tag_nodes[html_tag_nodes.length - 1]);
-	return new HtmlTag(html_tag_nodes.slice(1, html_tag_nodes.length - 1));
+	const claimed_nodes = html_tag_nodes.slice(1, html_tag_nodes.length - 1);
+	for (const n of claimed_nodes) {
+		n.claim_order = nodes.claim_info.total_claimed;
+		nodes.claim_info.total_claimed += 1;
+	}
+	return new HtmlTagHydration(claimed_nodes);
 }
 
 export function set_data(text, data) {
@@ -539,7 +631,7 @@ export function custom_event<T=any>(type: string, detail?: T, bubbles: boolean =
 }
 
 export function query_selector_all(selector: string, parent: HTMLElement = document.body) {
-	return Array.from(parent.querySelectorAll(selector));
+	return Array.from(parent.querySelectorAll(selector)) as ChildNodeArray;
 }
 
 export class HtmlTag {
@@ -547,27 +639,24 @@ export class HtmlTag {
 	e: HTMLElement;
 	// html tag nodes
 	n: ChildNode[];
-	// hydration claimed nodes
-	l: ChildNode[] | void;
 	// target
 	t: HTMLElement;
 	// anchor
 	a: HTMLElement;
 
-	constructor(claimed_nodes?: ChildNode[]) {
+	constructor() {
 		this.e = this.n = null;
-		this.l = claimed_nodes;
+	}
+
+	c(html: string) {
+		this.h(html);
 	}
 
 	m(html: string, target: HTMLElement, anchor: HTMLElement = null) {
 		if (!this.e) {
 			this.e = element(target.nodeName as keyof HTMLElementTagNameMap);
 			this.t = target;
-			if (this.l) {
-				this.n = this.l;
-			} else {
-				this.h(html);
-			}
+			this.c(html);
 		}
 
 		this.i(anchor);
@@ -592,6 +681,29 @@ export class HtmlTag {
 
 	d() {
 		this.n.forEach(detach);
+	}
+}
+
+export class HtmlTagHydration extends HtmlTag {
+	// hydration claimed nodes
+	l: ChildNode[] | void;
+
+	constructor(claimed_nodes?: ChildNode[]) {
+		super();
+		this.e = this.n = null;
+		this.l = claimed_nodes;
+	}
+	c(html: string) {
+		if (this.l) {
+			this.n = this.l;
+		} else {
+			super.c(html);
+		}
+	}
+	i(anchor) {
+		for (let i = 0; i < this.n.length; i += 1) {
+			insert_hydration(this.t, this.n[i], anchor);
+		}
 	}
 }