+++ test/parser/samples/element-with-attribute/input.svelte
@@ -1,0 +1,2 @@
+<span attr="foo"></span>
+<span attr='bar'></span>